Work trip model: shortest path
The code WorkModel.py aims to create the Relation matrix based on Euclidean distance so that the Shortest Path algorithm does not calculate shortest paths for points that are too far away from each other. There are 15 job purposes, which are treated separately (remove the points where they are = 0)

The code Shortest_Path.py creates a unique ID for each Points and relations layer and compute the shortest path between each set of points whose relation is in the relation layer.