feat: gather network fids by using a custom strategy
Use a secondary custom network cost strategy to save the fid of the
underlygin road network. This allows us to map the resulting graph
edges back onto the network without having to use geometry ops.

The current solution saves the integer ids as a comma separated
string. The plan is to not create a layer and instead gather
the results in a (numpy) table for easier processing.